+++ lib/ansible/modules/network/solace/solace_gather_facts.py
@@ -31,10 +31,12 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
                     'supported_by': 'community'}
 
 import ansible.module_utils.network.solace.solace_utils as su
+import ansible.module_utils.network.solace.solace_common as sc
 from ansible.module_utils.basic import AnsibleModule
 
 import traceback
 HAS_IMPORT_ERROR = False
+IMPORT_ERR_TRACEBACK = None
 try:
     from ansible.errors import AnsibleError
     import requests
@@ -42,6 +44,7 @@ try:
 except ImportError:
     HAS_IMPORT_ERROR = True
     IMPORT_ERR_TRACEBACK = traceback.format_exc()
+
 
 DOCUMENTATION = '''
 ---
@@ -140,10 +143,7 @@ ansible_facts.solace:
 class SolaceGatherFactsTask(su.SolaceTask):
 
     def __init__(self, module):
-        if HAS_IMPORT_ERROR:
-            exceptiondata = traceback.format_exc().splitlines()
-            exceptionarray = [exceptiondata[-1]] + exceptiondata[1:-1]
-            module.fail_json(msg="Missing module: %s" % exceptionarray[0], rc=1, exception=IMPORT_ERR_TRACEBACK)
+        sc.module_fail_on_import_error(module, HAS_IMPORT_ERROR, IMPORT_ERR_TRACEBACK)
         su.SolaceTask.__init__(self, module)
         return
 
@@ -171,7 +171,7 @@ class SolaceGatherFactsTask(su.SolaceTask):
         return True, about_info
 
     def _get_service_info_solace_cloud(self):
-        # https://api.solace.cloud/api/v0/services/{{serviceId}}
+        # GET https://api.solace.cloud/api/v0/services/{{serviceId}}
         path_array = [su.SOLACE_CLOUD_API_SERVICES_BASE_PATH, self.solace_config.solace_cloud_config['service_id']]
         return su.make_get_request(self.solace_config, path_array)
 
@@ -246,8 +246,8 @@ def make_get_request(solace_config, path_array):
                     headers={'x-broker-name': solace_config.x_broker},
                     params=None
         )
-        if su.ENABLE_LOGGING:
-            su.log_http_roundtrip(resp)
+        if sc.ENABLE_LOGGING:
+            sc.log_http_roundtrip(resp)
         if resp.status_code != 200:
             return False, su.parse_bad_response(resp), dict(resp.headers)
         return True, su.parse_good_response(resp), dict(resp.headers)
@@ -269,8 +269,8 @@ def make_sempv1_post_request(solace_config, xml_data):
                 headers=headers,
                 params=None
             )
-    if su.ENABLE_LOGGING:
-        su.log_http_roundtrip(resp)
+    if sc.ENABLE_LOGGING:
+        sc.log_http_roundtrip(resp)
     if resp.status_code != 200:
         raise AnsibleError("SEMP v1 call not successful. Pls check the log and raise an issue.")
     # SEMP v1 always returns 200 (it seems)