+++ test-test/tests-1-broker/solace_client_profile/run.sh
@@ -44,12 +44,13 @@ source $AS_TEST_HOME/lib/_run.env.sh $AS_TEST_RUNNER_ENV
     export ANSIBLE_SOLACE_ENABLE_LOGGING=true
     # select inventory
     export AS_TEST_BROKER_INVENTORY="$AS_TEST_HOME/lib/broker.inventories/local.broker.inventory.json"
-      # export AS_TEST_BROKER_INVENTORY=$(assertFile "$AS_TEST_HOME/lib/broker.inventories/cloud.broker.inventory.json") || exit
+    # export AS_TEST_BROKER_INVENTORY=$(assertFile "$AS_TEST_HOME/lib/broker.inventories/cloud.broker.inventory.json") || exit
     # select broker(s) inside inventory
     export AS_TEST_BROKERS="all"
 
     # playbook
     playbooks=(
+      "./update.playbook.yml"
       "./exceptions.playbook.yml"
       "./playbook.yml"
     )
@@ -79,15 +80,7 @@ for playbook in ${playbooks[@]}; do
                     --extra-vars "brokers=$AS_TEST_BROKERS" \
                     -vvv
 
-  if [[ $? != 0 ]]; then
-
-    echo "ERROR";
-    echo; echo "Show the log?"
-    echo; read -p 'Enter to continue, Ctrl-c to abort: ' continue; echo; echo
-
-    less $ANSIBLE_SOLACE_LOG_FILE
-
-  fi
+  if [[ $? != 0 ]]; then echo ">>> ERR: $AS_TEST_SCRIPT_PATH"; echo; exit 1; fi
 
 done