+++ test-test/tests-solace-cloud/solace_cloud_service/run.sh
@@ -40,13 +40,31 @@ source $AS_TEST_HOME/lib/_run.env.sh $AS_TEST_RUNNER_ENV
 
   ############################################################################################################################
   # SELECT
-    # logging
-    export ANSIBLE_SOLACE_ENABLE_LOGGING=true
+    # logging & debug: ansible
+      # ansibleLogFile="./tmp/ansible.log"
+      # export ANSIBLE_LOG_PATH="$ansibleLogFile"
+      # export ANSIBLE_DEBUG=False
+      export ANSIBLE_VERBOSITY=3
+    # logging: ansible-solace
+      export ANSIBLE_SOLACE_LOG_PATH="./tmp/ansible-solace.log"
+      export ANSIBLE_SOLACE_ENABLE_LOGGING=True
     # select inventory
-    export AS_TEST_BROKER_INVENTORY="$AS_TEST_HOME/lib/broker.inventories/local.broker.inventory.json"
-    #export AS_TEST_BROKER_INVENTORY=$(assertFile "$AS_TEST_HOME/lib/broker.inventories/cloud.broker.inventory.json") || exit
-    # select broker(s) inside inventory
-    export AS_TEST_BROKERS="all"
+      solaceCloudAccountsInventoryFile=$(assertFile "$AS_TEST_SCRIPT_PATH/lib/solace-cloud-accounts.inventory.yml") || exit
+    # select account(s) inside inventory
+      solaceCloudAccounts="all"
+
+    # playbook
+    playbooks=(
+      "./create.playbook.yml"
+      "./create.playbook.yml" # run it twice, to test idempotency
+      "./facts.playbook.yml"
+      "./ex_1.playbook.yml"
+      "./ex_2.playbook.yml"
+      "./ex_3.playbook.yml"
+      "./delete.playbook.yml"
+      "./delete.playbook.yml" # run it twice, to test idempotency
+    )
+
   # END SELECT
 
 
@@ -56,24 +74,21 @@ x=$(wait4Key)
 ##############################################################################################################################
 # Prepare
 
-ANSIBLE_SOLACE_LOG_FILE="$AS_TEST_SCRIPT_PATH/ansible-solace.log"
-rm -f $ANSIBLE_SOLACE_LOG_FILE
-
-$AS_TEST_HOME/tests-embeddable/wait-until-broker-available/_run.call.sh $AS_TEST_BROKER_INVENTORY
-if [[ $? != 0 ]]; then echo "ERR >>> aborting."; echo; exit 1; fi
+mkdir ./tmp > /dev/null 2>&1
+rm -f ./tmp/*.*
 
 ##############################################################################################################################
 # Run
+for playbook in ${playbooks[@]}; do
 
-playbook="./playbook.yml"
+  ansible-playbook \
+                    -i $solaceCloudAccountsInventoryFile \
+                    $playbook \
+                    --extra-vars "SOLACE_CLOUD_ACCOUNTS=$solaceCloudAccounts"
 
-# --step --check -vvv
-ansible-playbook -i $AS_TEST_BROKER_INVENTORY \
-                  $playbook \
-                  --extra-vars "brokers=$AS_TEST_BROKERS" \
-                  -vvv
-
-if [[ $? != 0 ]]; then echo ">>> ERR: $AS_TEST_SCRIPT_PATH"; echo; exit 1; fi
+  if [[ $? != 0 ]]; then echo ">>> ERR: $AS_TEST_SCRIPT_PATH"; echo; exit 1; fi
+  
+done
 
 ###
 # The End.